requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
cffi==1.13.2
chardet==3.0.4
discord.py==1.2.5
ffmpeg==1.4
idna==2.8
multidict==4.7.3
pycparser==2.19
PyNaCl==1.3.0
six==1.13.0
websockets==6.0
yarl==1.4.2